ui improvements to react structure
