+++ components/common/Typography/index.ts
@@ -1,2 +1,3 @@
 export * from "./Text";
 export * from "./Heading";
+export * from "./LinkText";
