+++ components/icons/LinkIcon.tsx
@@ -1,0 +1,26 @@
+export function LinkIcon(props: JSX.IntrinsicElements["svg"]) {
+  return (
+    <svg
+      width="12"
+      height="12"
+      viewBox="0 0 12 12"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <g clip-path="url(#clip0_8_87418)">
+        <path
+          fill-rule="evenodd"
+          clip-rule="evenodd"
+          d="M8.68934 2.25H6V0.75H11.25V6H9.75V3.31066L5.03033 8.03033L3.96967 6.96967L8.68934 2.25ZM0.75 2.25H3.75V3.75H2.25V9.75H8.25V8.25H9.75V11.25H0.75V2.25Z"
+          fill="currentColor"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_8_87418">
+          <rect width="12" height="12" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
