+++ components/molecules/index.ts
@@ -1,0 +1,3 @@
+export * from "./FormationSection";
+export * from "./TeamInfoSection";
+export * from "./JerseyDetailsSection";
